feat :[#68] 마일스톤 추가 버튼 View 구현
