--- iOS/issue-tracker/issue-tracker/Main/Milestone/MilestoneViewController.swift
@@ -1,26 +1,0 @@
-//
-//  MilestoneViewController.swift
-//  issue-tracker
-//
-//  Created by Song on 2021/06/09.
-//
-
-import UIKit
-
-class MilestoneViewController: UIViewController {
-    
-    private var loginInfo: LoginInfo?
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = UIColor.yellow
-        title = "마일스톤"
-    }
-    
-}
-
-extension MilestoneViewController: LoginInfoContainer {
-    func setup(loginInfo: LoginInfo) {
-        self.loginInfo = loginInfo
-    }
-}
